updated vacancies.models, updated fill_test_db, updated search.models
